Created programs/ dir
effects.cpp was getting too long. Created programs-common.h and .cpp

Transition to FastLED() is complete, save for some tweaks and
adaptation.
