+++ Code/Flamingo Rev B/platformio.ini
@@ -18,7 +18,7 @@ lib_deps =
   FastLED
   I2Cdevlib-Core
   I2Cdevlib-ADXL345
-  I2Cdevlib-MPU6050
+;  I2Cdevlib-MPU6050
 
 ;[env:protrinket5]
 ;platform = atmelavr
